chore: refactor unit tests and rename function
